+++ src/main/java/io/micrometer/NewRelicRegistryConfig.java
@@ -1,8 +1,6 @@
 /*
- * ---------------------------------------------------------------------------------------------
- *  Copyright (c) 2019 New Relic Corporation. All rights reserved.
- *  Licensed under the Apache 2.0 License. See LICENSE in the project root directory for license information.
- * --------------------------------------------------------------------------------------------
+ * Copyright 2020 New Relic Corporation. All rights reserved.
+ * SPDX-License-Identifier: Apache-2.0
  */
 
 package io.micrometer;
@@ -11,6 +9,7 @@ import io.micrometer.core.instrument.config.MissingRequiredConfigurationExceptio
 import io.micrometer.core.instrument.step.StepRegistryConfig;
 
 public interface NewRelicRegistryConfig extends StepRegistryConfig {
+
   @Override
   default String prefix() {
     return "newrelic";
@@ -18,9 +17,10 @@ public interface NewRelicRegistryConfig extends StepRegistryConfig {
 
   default String apiKey() {
     String apiKey = get(prefix() + ".apiKey");
-    if (apiKey == null)
+    if (apiKey == null) {
       throw new MissingRequiredConfigurationException(
           "apiKey must be set to report metrics to New Relic");
+    }
     return apiKey;
   }
 
@@ -30,5 +30,26 @@ public interface NewRelicRegistryConfig extends StepRegistryConfig {
    */
   default String uri() {
     return get(prefix() + ".uri");
+  }
+
+  /**
+   * Return the service name which this registry will report as. Maps to the "service.name"
+   * attribute on the metrics.
+   *
+   * @return The Service Name.
+   */
+  default String serviceName() {
+    return null;
+  }
+
+  /**
+   * Turn on "audit mode" in the underlying New Relic Telemetry SDK. This will log all data sent to
+   * the New Relic APIs. Be aware that if there is sensitive information in the data being sent that
+   * it will be sent to wherever the Telemetry SDK logs are configured to go.
+   *
+   * @return true if audit mode should be enabled.
+   */
+  default boolean enableAuditMode() {
+    return false;
   }
 }